Refactor: Separate out Zephyr setup functionality and add to new Zephyr Project
